+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -1934,6 +1934,22 @@ dependency to version 2.1.16.
 - [LPS-73472]: Update the [Liferay Gradle Plugins JS Transpiler] dependency to
 version 2.3.16.
 
+## 3.4.7 - 2017-07-18
+
+### Changed
+- [LPS-72705]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.0.36.
+- [LPS-72705]: Update the [Liferay Source Formatter] dependency to version
+1.0.425.
+
+## 3.4.8 - 2017-07-19
+
+### Changed
+- [LPS-72705]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.0.37.
+- [LPS-72705]: Update the [Liferay Source Formatter] dependency to version
+1.0.426.
+
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Alloy Taglib]: https://github.com/liferay/alloy-taglibs
 [Liferay Ant BND]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/ant-bnd